sp-wdio:
Swapped the order of commands (making the command for w3c
first and the JSON wire protocol 2nd) since the generated code is unconditionally
calling `getSession` in some cases.

```
async function getCapabilities(browser) {
    var _a, _b;
    return (_b = (_a = browser.getSession) === null || _a === void 0 ? void 0 : _a.call(browser)) !== null && _b !== void 0 ? _b : browser.capabilities;
}
```

This causes tests to error when running on
Chrome with sensible defaults (e.g., w3c true), which is the common
configuration. The specific case that spurred this fix
was from running a coverage test on mobile web (Chrome on Android) on
Sauce Labs.