Add better support for Eyes.abort (#798)
* core: add test to check for abort in closeAllEyes

* core: add abort support to closeAllEyes

* Updated abort so it fails gracefully when a valid eyes instance is not
present.

Pulled the code for abort into its own function so it could be unit
tested since there was no pre-existing test harness to work with the
server through sockets in an integration test. I started down the path to add
this kind of test but felt like it was just out of reach and
taking too long. I think this kind of test is worth doing. It just requires
more time. I've commited the test for it, marked as skipped
for now.

* uni: update changelog

* core e2e test verify isAborted on test results

* change app name in test

* add aborted tests to runner

* fix universal support for abort after close

Co-authored-by: Amit Zur <amit.zur@applitools.com>