test-utils:
Fixed waitForDockerBrowsers so it throws when the remoteUrl is not
provided. This will make it easy to know when a workflow in GitHub
Actions does not have the CVG_TESTS_REMOTE environment variable configured.

Also refactored waitForDockerBrowser so it is unaware of env vars. Pushed that
responsibility to the caller and verified e2e by linking in an SDK and
running a coverage test with and without the necessary environment variables set.

Included a mention of the environment variable CVG_TESTS_REMOTE in the
error message when the remoteUrl is not provided to
waitForDockerBrowsers.