bongo: fix verify-installed-versions
In newer versions of npm (> 7, it seems) they changed the file format
(e.g., lockFileVersion and this doesn't play well with packing/installing
a package locally and running `npm ls`.

We have had the ability to inspect the package-lock.json for some time,
but it was switched back to using `npm ls` due to issues a long time ago
(I don't recall exactly why), but this commit tests a package-lock.json file
with the new (current) file format and fixes the issues experienced at
this time.