universal e2e tests client (#1078)
In order to test the communication between the `eyes-universal` and other platform (like `Python` or `Ruby` for example) this `client` will test the communication between `eyes-universal` (a.k.a. server) and a `client` (a.k.a. `nodejs` client).

This change is here to replace [this close](https://github.com/applitools/eyes.sdk.javascript1/pull/1010) pr that try to make this `client` as a standalone package.
 
Co-authored-by: Amit Zur <amit.zur@applitools.com>